refactoring code with flake8
