+++ Pipfile
@@ -14,6 +14,7 @@ pytest = "*"
 coverage = "*"
 pytest-cov = "*"
 locust = "*"
+flake8 = "*"
 
 [dev-packages]
