+++ gudlft/db.py
@@ -1,0 +1,13 @@
+import json
+
+
+def loadClubs():
+    with open('gudlft/json/clubs.json') as c:
+        listOfClubs = json.load(c)['clubs']
+        return listOfClubs
+
+
+def loadCompetitions():
+    with open('gudlft/json/competitions.json') as comps:
+        listOfCompetitions = json.load(comps)['competitions']
+        return listOfCompetitions
