+++ setup.cfg
@@ -1,0 +1,10 @@
+[flake8]
+exclude =
+    .gitignore,
+    .pytest_cache,
+    .venv,
+    htmlcov
+max-complexity = 10
+
+[tool:pytest]
+testpaths = tests
